requirements.txt
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.2
python-dotenv==1.0.1
requests
playwright
